+++ src/module/personal/savedMessage/SavedMessage.controller.ts
@@ -3,33 +3,43 @@ import {SavedMessageService} from './SavedMessage.service';
 import Json from "../../../util/ReturnJson";
 import Response from "../../../util/Response";
 import {SavedMessage} from "../../base/SavedMessage";
+import PromiseVerify from "../../base/PromiseVerify";
 
 @Controller()
 export class SavedMessageController extends SavedMessage {
     constructor(private readonly appService: SavedMessageService) {
         super();
-        this.init();
     }
 
     @Post()
     async createSavedMessage() {
-        this.verifySavedMessage().then(async () => {
+        this.init();
 
-            await this.appService.addSavedMessage(this.phoneNumber);
+        let haveErr = await PromiseVerify.all([
+            this.verifySavedMessage()
+        ]);
 
-            return Json.builder(Response.HTTP_CREATED);
+        if (haveErr)
+            return haveErr;
 
-        });
+        await this.appService.addSavedMessage(this.phoneNumber);
+
+        return Json.builder(Response.HTTP_CREATED);
     }
 
     @Delete()
     async deleteSavedMessage() {
-        this.verifySavedMessage().then(async () => {
+        this.init();
 
-            await this.appService.removeSavedMessage(this.phoneNumber);
+        let haveErr = await PromiseVerify.all([
+            this.verifySavedMessage()
+        ]);
 
-            return Json.builder(Response.HTTP_OK);
+        if (haveErr)
+            return haveErr;
 
-        });
+        await this.appService.removeSavedMessage(this.phoneNumber);
+
+        return Json.builder(Response.HTTP_OK);
     }
 }
